fix(bench): close remaining BF16-K/FP16-V contamination in 3 benches
context_softmax_attention_bench, sliding_window_attention_bench, and
kv_cache_append_bench still hardcoded DType::FP16 for the V plane,
the same asymmetric-storage contamination fixed in the production and
test code in the prior commit. context/sliding-window benches
crashed outright against the now-symmetric-BF16 validation
("context K/V must be BF16") until fixed. Confirmed via benchmark
comparison against a clean master-baseline worktree build: BF16
kv_cache_append and context/sliding-window attention throughput is
within noise of the pre-merge baseline. The causal small_t BF16
decode path benchmarks ~15-20% slower than baseline, which is the
expected, deliberate cost of neroued's FP32 (vs BF16) split-reduction
partial-accumulator precision improvement kept in the prior commit,
not a regression.

+++ bench/ops/kv_cache_append_bench.cu
@@ -362,7 +362,7 @@ PagedKVBatchLayerView make_prefix_paged_view(DeviceBuffer& k, DeviceBuffer& v,
             k.p, DType::BF16,
             {kPrefixHeadDim, kPagedKVPageSize, kRingCapacity / kPagedKVPageSize, kPrefixKvHeads}),
         .v_pages = Tensor(
-            v.p, DType::FP16,
+            v.p, DType::BF16,
             {kPrefixHeadDim, kPagedKVPageSize, kRingCapacity / kPagedKVPageSize, kPrefixKvHeads}),
         .block_tables = Tensor(block_tables.p, DType::I32, {kRingCapacity / kPagedKVPageSize, 1}),
         .head_dim     = kPrefixHeadDim,
@@ -374,7 +374,7 @@ PagedKVBatchLayerView make_prefix_paged_view(DeviceBuffer& k, DeviceBuffer& v,
 CyclicKVCacheLayerView make_prefix_cyclic_view(DeviceBuffer& k, DeviceBuffer& v) {
     return {
         .k        = Tensor(k.p, DType::BF16, {kPrefixHeadDim, kRingCapacity, kPrefixKvHeads, 1}),
-        .v        = Tensor(v.p, DType::FP16, {kPrefixHeadDim, kRingCapacity, kPrefixKvHeads, 1}),
+        .v        = Tensor(v.p, DType::BF16, {kPrefixHeadDim, kRingCapacity, kPrefixKvHeads, 1}),
         .capacity = kRingCapacity,
         .padded_capacity = kRingCapacity,
         .num_kv_heads    = kPrefixKvHeads,
